🛂 Add authorization

+++ src/api/notes/handler.js
@@ -16,8 +16,10 @@ module.exports = class NoteHandler {
     try {
       this.validator.validateNotePayload(request.payload);
       const { title = 'untitled', body, tags } = request.payload;
-
-      const noteId = await this.service.addNote({ title, body, tags });
+      const { id: credentialId } = request.auth.credentials;
+      const noteId = await this.service.addNote({
+        title, body, tags, owner: credentialId,
+      });
 
       const response = h.response({
         status: 'success',
@@ -49,8 +51,9 @@ module.exports = class NoteHandler {
     }
   }
 
-  async getNotesHandler() {
-    const notes = await this.service.getNotes();
+  async getNotesHandler(request) {
+    const { id: credentialId } = request.auth.credentials;
+    const notes = await this.service.getNotes(credentialId);
     return {
       status: 'success',
       data: {
@@ -62,6 +65,9 @@ module.exports = class NoteHandler {
   async getNoteByIdHandler(request, h) {
     try {
       const { id } = request.params;
+      const { id: credentialId } = request.auth.credentials;
+
+      await this.service.verifyNoteOwner(id, credentialId);
       const note = await this.service.getNoteById(id);
       return {
         status: 'success',
@@ -94,6 +100,9 @@ module.exports = class NoteHandler {
     try {
       this.validator.validateNotePayload(request.payload);
       const { id } = request.params;
+      const { id: credentialId } = request.auth.credentials;
+
+      await this.service.verifyNoteOwner(id, credentialId);
       await this.service.editNoteById(id, request.payload);
       const response = h.response({
         status: 'success',
@@ -125,6 +134,9 @@ module.exports = class NoteHandler {
   async deleteNoteByIdHandler(request, h) {
     try {
       const { id } = request.params;
+      const { id: credentialId } = request.auth.credentials;
+
+      await this.service.verifyNoteOwner(id, credentialId);
       await this.service.deleteNoteById(id);
       const response = h.response({
         status: 'success',
